formating

+++ cpp/FESpace/operationTypeFE.hpp
@@ -36,51 +36,51 @@ CutFEM-Library. If not, see <https://www.gnu.org/licenses/>
 #define _OPERATIONTYPEFE_HPP
 
 enum operatortype {
-   op_id    = 0,
-   op_dx    = 1,
-   op_dy    = 2,
-   op_dz    = 3,
-   op_dxx   = 4,
-   op_dyy   = 5,
-   op_dzz   = 6,
-   op_dyx   = 7,
-   op_dxy   = 7,
-   op_dzx   = 8,
-   op_dxz   = 8,
-   op_dzy   = 9,
-   op_dyz   = 9,
-   op_all   = 10,
-   op_Dall  = 10,
-   // need to change those value, do -1, start from 10
-   op_dxxx  = 11,
-   op_dxxy  = 14,
-   op_dxxz  = 15,
-   op_dxyx  = 14,
-   op_dxyy  = 16,
-   op_dxyz  = 17,
-   op_dxzx  = 15,
-   op_dxzy  = 17,
-   op_dxzz  = 18,
-   op_dyxx  = 14,
-   op_dyxy  = 16,
-   op_dyxz  = 17,
-   op_dyyx  = 16,
-   op_dyyy  = 12,
-   op_dyyz  = 19,
-   op_dyzx  = 17,
-   op_dyzy  = 19,
-   op_dyzz  = 20,
-   op_dzxx  = 15,
-   op_dzxy  = 17,
-   op_dzxz  = 18,
-   op_dzyx  = 17,
-   op_dzyy  = 19,
-   op_dzyz  = 20,
-   op_dzzx  = 18,
-   op_dzzy  = 20,
-   op_dzzz  = 13,
-   op_DDall = 21,
-   op_All   = 21
+    op_id    = 0,
+    op_dx    = 1,
+    op_dy    = 2,
+    op_dz    = 3,
+    op_dxx   = 4,
+    op_dyy   = 5,
+    op_dzz   = 6,
+    op_dyx   = 7,
+    op_dxy   = 7,
+    op_dzx   = 8,
+    op_dxz   = 8,
+    op_dzy   = 9,
+    op_dyz   = 9,
+    op_all   = 10,
+    op_Dall  = 10,
+    // need to change those value, do -1, start from 10
+    op_dxxx  = 11,
+    op_dxxy  = 14,
+    op_dxxz  = 15,
+    op_dxyx  = 14,
+    op_dxyy  = 16,
+    op_dxyz  = 17,
+    op_dxzx  = 15,
+    op_dxzy  = 17,
+    op_dxzz  = 18,
+    op_dyxx  = 14,
+    op_dyxy  = 16,
+    op_dyxz  = 17,
+    op_dyyx  = 16,
+    op_dyyy  = 12,
+    op_dyyz  = 19,
+    op_dyzx  = 17,
+    op_dyzy  = 19,
+    op_dyzz  = 20,
+    op_dzxx  = 15,
+    op_dzxy  = 17,
+    op_dzxz  = 18,
+    op_dzyx  = 17,
+    op_dzyy  = 19,
+    op_dzyz  = 20,
+    op_dzzx  = 18,
+    op_dzzy  = 20,
+    op_dzzz  = 13,
+    op_DDall = 21,
+    op_All   = 21
 };
 
 typedef unsigned int What_d;
@@ -112,23 +112,23 @@ const unsigned int Fop_DAll = 20;
 
 inline What_d Fwhatd(const operatortype op) { return 1 << op; }
 inline What_d Fwhatd(const int op) {
-   if (op == 1)
-      return Fop_D0;
-   else if (op < 5)
-      return Fop_D0 | Fop_D1;
-   else
-      return Fop_Dall;
+    if (op == 1)
+        return Fop_D0;
+    else if (op < 5)
+        return Fop_D0 | Fop_D1;
+    else
+        return Fop_Dall;
 }
 
 inline int getLastop(int i, int j) {
-   double lastop = std::max(i, j);
-   if (lastop == 0)
-      return lastop + 1;
-   if (lastop <= op_dz)
-      return op_dz + 1;
-   if (lastop <= op_dyz)
-      return op_Dall;
-   return 0;
+    int lastop = std::max(i, j);
+    if (lastop == 0)
+        return lastop + 1;
+    if (lastop <= op_dz)
+        return op_dz + 1;
+    if (lastop <= op_dyz)
+        return op_Dall;
+    return 0;
 }
 
 const int last_operatortype                     = 20;
@@ -138,9 +138,9 @@ const bool operatortypeValue[last_operatortype] = {
 };
 
 inline void initwhatd(bool *whatd, int k) {
-   for (int i = 0; i < last_operatortype; i++)
-      whatd[i] = false;
-   whatd[k] = true;
+    for (int i = 0; i < last_operatortype; i++)
+        whatd[i] = false;
+    whatd[k] = true;
 }
 
 #endif
